skills/agent-browser-bridge/main.py
#!/usr/bin/env python3
"""
Agent Browser Bridge Service
基于 browser-use 的 OpenClaw 浏览器桥接服务
"""

import asyncio
import base64
import json
import os
from datetime import datetime
from typing import Optional

from dotenv import load_dotenv
from fastapi import FastAPI, WebSocket, WebSocketDisconnect
from fastapi.middleware.cors import CORSMiddleware

# Load environment variables
load_dotenv()

# Placeholder imports - will be replaced with actual browser-use imports
# from browser_use import Agent, Browser, ChatOpenAI

app = FastAPI(title="Agent Browser Bridge", version="1.0.0")

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


class BrowserSession:
    """管理浏览器会话"""
    
    def __init__(self):
        self.browser = None
        self.agent = None
        self.current_task = None
        self.is_running = False
        self.websocket: Optional[WebSocket] = None
    
    async def initialize(self):
        """初始化浏览器"""
        try:
            # TODO: 集成 browser-use
            # self.browser = Browser(
            #     headless=os.getenv("BROWSER_HEADLESS", "false").lower() == "true"
            # )
            print("🌐 Browser initialized (placeholder)")
        except Exception as e:
            print(f"❌ Failed to initialize browser: {e}")
            raise
    
    async def execute_task(self, task: str, websocket: WebSocket):
        """执行任务并实时推送状态"""
        self.websocket = websocket
        self.is_running = True
        self.current_task = task
        
        try:
            # TODO: 集成 browser-use Agent
            # self.agent = Agent(
            #     task=task,
            #     llm=ChatOpenAI(model="gpt-4o"),
            #     browser=self.browser,
            # )
            
            # 模拟执行过程
            steps = [
                {"action": "goto", "url": "https://google.com", "description": "Navigating to Google"},
                {"action": "click", "selector": "[name='q']", "description": "Clicking search box"},
                {"action": "type", "text": task, "description": f"Typing: {task}"},
                {"action": "press", "key": "Enter", "description": "Pressing Enter"},
            ]
            
            for step in steps:
                if not self.is_running:
                    break
                
                # 发送状态更新
                await self._send_state_update(step)
                
                # 检查是否需要人工接管
                if await self._check_handoff(step):
                    handoff_result = await self._request_handoff(step)
                    if not handoff_result:
                        await self._send_message({
                            "type": "task_cancelled",
                            "reason": "human_cancelled"
                        })
                        return
                
                await asyncio.sleep(1)  # 模拟执行时间
            
            # 任务完成
            await self._send_message({
                "type": "task_completed",
                "result": {
                    "task": task,
                    "completed_at": datetime.now().isoformat(),
                    "summary": f"Task '{task}' completed successfully"
                }
            })
            
        except Exception as e:
            await self._send_message({
                "type": "error",
                "message": str(e)
            })
        finally:
            self.is_running = False
    
    async def _send_state_update(self, step: dict):
        """发送状态更新"""
        # TODO: 实际截图
        # screenshot = await self.browser.take_screenshot()
        # screenshot_b64 = base64.b64encode(screenshot).decode()
        
        placeholder_screenshot = self._create_placeholder_screenshot()
        
        await self._send_message({
            "type": "state_update",
            "url": step.get("url", "https://example.com"),
            "title": "Example Page",
            "screenshot": placeholder_screenshot,
            "action": step["action"],
            "description": step.get("description", ""),
            "timestamp": datetime.now().isoformat()
        })
    
    async def _check_handoff(self, step: dict) -> bool:
        """检查是否需要人工接管"""
        # TODO: 实现登录页面检测
        # 检测 URL、页面元素等
        return step["action"] == "goto" and "login" in step.get("url", "").lower()
    
    async def _request_handoff(self, step: dict) -> bool:
        """请求人工接管"""
        await self._send_message({
            "type": "handoff_required",
            "reason": "login_page_detected",
            "url": step.get("url", ""),
            "message": "检测到登录页面，请人工完成登录后继续"
        })
        
        # 等待客户端响应（超时 5 分钟）
        try:
            response = await asyncio.wait_for(
                self._wait_for_handoff_response(),
                timeout=300
            )
            return response.get("success", False)
        except asyncio.TimeoutError:
            await self._send_message({
                "type": "handoff_timeout",
                "message": "人工接管超时"
            })
            return False
    
    async def _wait_for_handoff_response(self) -> dict:
        """等待人工接管响应"""
        # 实际的响应会在 WebSocket 消息处理中设置
        future = asyncio.Future()
        self._handoff_future = future
        return await future
    
    async def _send_message(self, message: dict):
        """发送消息到客户端"""
        if self.websocket:
            await self.websocket.send_json(message)
    
    def _create_placeholder_screenshot(self) -> str:
        """创建占位截图"""
        # TODO: 实际实现
        return "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg=="
    
    async def stop(self):
        """停止当前任务"""
        self.is_running = False
    
    async def close(self):
        """关闭浏览器"""
        # TODO: 关闭 browser-use browser
        # if self.browser:
        #     await self.browser.close()
        print("🌐 Browser closed")


# 全局会话管理
sessions: dict[str, BrowserSession] = {}


@app.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):
    """WebSocket 主入口"""
    await websocket.accept()
    session_id = str(id(websocket))
    
    print(f"🔌 Client connected: {session_id}")
    
    # 创建新会话
    session = BrowserSession()
    sessions[session_id] = session
    
    try:
        await session.initialize()
        
        while True:
            # 接收客户端消息
            data = await websocket.receive_text()
            message = json.loads(data)
            
            msg_type = message.get("type")
            
            if msg_type == "execute_task":
                task = message.get("task", "")
                options = message.get("options", {})
                
                # 异步执行任务
                asyncio.create_task(
                    session.execute_task(task, websocket)
                )
            
            elif msg_type == "get_state":
                # 获取当前状态
                await session._send_message({
                    "type": "state",
                    "url": "https://example.com",
                    "title": "Example",
                    "screenshot": session._create_placeholder_screenshot()
                })
            
            elif msg_type == "human_completed":
                # 人工接管完成
                if hasattr(session, '_handoff_future') and session._handoff_future:
                    session._handoff_future.set_result({
                        "success": message.get("success", True)
                    })
            
            elif msg_type == "stop":
                await session.stop()
                await session._send_message({
                    "type": "stopped"
                })
    
    except WebSocketDisconnect:
        print(f"🔌 Client disconnected: {session_id}")
    except Exception as e:
        print(f"❌ Error: {e}")
        await websocket.send_json({
            "type": "error",
            "message": str(e)
        })
    finally:
        await session.close()
        del sessions[session_id]


@app.get("/health")
async def health_check():
    """健康检查"""
    return {
        "status": "ok",
        "version": "1.0.0",
        "active_sessions": len(sessions)
    }


@app.get("/")
async def root():
    """根路径"""
    return {
        "service": "Agent Browser Bridge",
        "version": "1.0.0",
        "status": "running",
        "websocket": "ws://localhost:8765/ws"
    }


if __name__ == "__main__":
    import uvicorn
    
    host = os.getenv("HOST", "0.0.0.0")
    port = int(os.getenv("PORT", 8765))
    
    print(f"🚀 Starting Agent Browser Bridge on {host}:{port}")
    print(f"📡 WebSocket: ws://{host}:{port}/ws")
    
    uvicorn.run(app, host=host, port=port)
