skills/agent-browser-bridge/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
browser-use>=0.1.0
playwright>=1.49.0
pydantic>=2.10.0
python-dotenv>=1.0.0
Pillow>=11.0.0
